+++ lib/src/cache_stream/cache_downloader/download_response_listener.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import '../../../http_cache_stream.dart';
 import '../../etc/chunked_bytes_buffer.dart';
 import '../../etc/timeout_timer.dart';
+import '../../models/config/stream_cache_config.dart';
 import '../../models/exceptions/http_exceptions.dart';
 
 class DownloadResponseListener {