+++ lib/src/cache_stream/cache_downloader/downloader.dart
@@ -1,11 +1,13 @@
 import 'dart:async';
 
-import 'package:http_cache_stream/http_cache_stream.dart';
 import 'package:http_cache_stream/src/cache_stream/response_streams/download_stream.dart';
 import 'package:http_cache_stream/src/models/exceptions/invalid_cache_exceptions.dart';
 
 import '../../etc/pause_counter.dart';
+import '../../models/config/stream_cache_config.dart';
 import '../../models/exceptions/http_exceptions.dart';
+import '../../models/metadata/cached_response_headers.dart';
+import '../../models/stream_requests/int_range.dart';
 import 'download_response_listener.dart';
 
 class Downloader {