+++ lib/src/cache_stream/response_streams/combined_data_stream.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 
-import '../../../http_cache_stream.dart';
 import '../../etc/extensions/stream_extensions.dart';
+import '../../models/cache_files/cache_files.dart';
+import '../../models/config/stream_cache_config.dart';
 import '../../models/exceptions/stream_response_exceptions.dart';
+import '../../models/stream_requests/int_range.dart';
 import '../../models/stream_response/stream_response_range.dart';
 import 'buffered_data_stream.dart';
 import 'cache_file_stream.dart';