+++ lib/src/cache_stream/response_streams/download_stream.dart
@@ -3,11 +3,7 @@ import 'dart:io';
 
 import 'package:dio/dio.dart' as libdio;
 import 'package:http_cache_stream/http_cache_stream.dart';
-import 'package:http_cache_stream/src/models/exceptions/invalid_cache_exceptions.dart';
-import 'package:http_cache_stream/src/models/http_range/http_range_response.dart';
-import 'package:util_xx/Httpxx.dart';
-
-import '../../models/exceptions/http_exceptions.dart';
+import 'package:util_xx/util_xx.dart';
 
 class DownloadStream extends Stream<List<int>> {
   final libdio.ResponseBody _streamedResponse;