+++ lib/src/models/stream_response/cache_download_stream_response.dart
@@ -1,9 +1,11 @@
 import 'dart:async';
 
-import 'package:http_cache_stream/http_cache_stream.dart';
-
 import '../../cache_stream/response_streams/buffered_data_stream.dart';
+import '../config/stream_cache_config.dart';
 import '../exceptions/stream_response_exceptions.dart';
+import '../metadata/cached_response_headers.dart';
+import '../stream_requests/int_range.dart';
+import 'stream_response.dart';
 import 'stream_response_range.dart';
 
 /// A stream response that buffers data from the cache download stream and serves it according to the specified range.