+++ lib/src/models/stream_response/range_download_stream_response.dart
@@ -1,9 +1,10 @@
 import 'dart:async';
 
-import 'package:http_cache_stream/http_cache_stream.dart';
-
 import '../../cache_stream/response_streams/download_stream.dart';
 import '../../etc/chunked_bytes_buffer.dart';
+import '../config/stream_cache_config.dart';
+import '../stream_requests/int_range.dart';
+import 'stream_response.dart';
 
 class RangeDownloadStreamResponse extends StreamResponse {
   final DownloadStream _downloadStream;