+++ lib/src/request_handler/request_handler.dart
@@ -6,10 +6,6 @@ import 'package:util_xx/util_xx.dart';
 
 import '../../http_cache_stream.dart';
 import '../etc/mime_types.dart';
-import '../models/exceptions/invalid_cache_exceptions.dart';
-import '../models/http_range/http_range.dart';
-import '../models/http_range/http_range_request.dart';
-import '../models/http_range/http_range_response.dart';
 import '../request_handler/socket_handler.dart';
 
 class RequestHandler {